+++ src/columns/start_time.rs
@@ -2,9 +2,16 @@ use crate::process::ProcessInfo;
 use crate::{column_default, Column};
 #[cfg(not(target_os = "windows"))]
 use chrono::offset::TimeZone;
-use chrono::{DateTime, Local};
+use chrono::{self, DateTime, Duration, Local};
+#[cfg(target_os = "linux")]
+use lazy_static::lazy_static;
 use std::cmp;
 use std::collections::HashMap;
+
+#[cfg(target_os = "linux")]
+lazy_static! {
+    static ref TICKS_PER_SECOND: i64 = { procfs::ticks_per_second().unwrap() };
+}
 
 pub struct StartTime {
     header: String,
@@ -12,6 +19,8 @@ pub struct StartTime {
     fmt_contents: HashMap<i32, String>,
     raw_contents: HashMap<i32, DateTime<Local>>,
     width: usize,
+    #[cfg(target_os = "linux")]
+    boot_time: DateTime<Local>,
 }
 
 impl StartTime {
@@ -24,6 +33,8 @@ impl StartTime {
             width: 0,
             header,
             unit,
+            #[cfg(target_os = "linux")]
+            boot_time: procfs::boot_time().unwrap_or(Local.timestamp(0, 0)),
         }
     }
 }
@@ -31,13 +42,11 @@ impl StartTime {
 #[cfg(target_os = "linux")]
 impl Column for StartTime {
     fn add(&mut self, proc: &ProcessInfo) {
-        let start_time = proc
-            .curr_proc
-            .stat
-            .starttime()
-            .unwrap_or(Local.timestamp(0, 0));
-        let raw_content = start_time;
-        let fmt_content = format!("{}", start_time.format("%Y/%m/%d %H:%M"));
+        let starttime = proc.curr_proc.stat.starttime;
+        let seconds_since_boot = starttime as f32 / *TICKS_PER_SECOND as f32;
+        let raw_content =
+            self.boot_time + Duration::milliseconds((seconds_since_boot * 1000.0) as i64);
+        let fmt_content = format!("{}", raw_content.format("%Y/%m/%d %H:%M"));
 
         self.fmt_contents.insert(proc.pid, fmt_content);
         self.raw_contents.insert(proc.pid, raw_content);